Conflicts:
	engines/mohawk/console.cpp
	engines/mohawk/console.h
	engines/mohawk/detection.cpp
	engines/mohawk/graphics.cpp
	engines/mohawk/graphics.h
	engines/mohawk/module.mk

+++ engines/gob/scenery.cpp
@@ -18,9 +18,6 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
- * $URL$
- * $Id$
- *
  */
 
 #include "common/endian.h"
@@ -625,7 +622,7 @@ void Scenery::updateAnim(int16 layer, int16 frame, int16 animation, int16 flags,
 		if (frame >= (int32)_vm->_vidPlayer->getFrameCount(obj.videoSlot - 1))
 			frame = _vm->_vidPlayer->getFrameCount(obj.videoSlot - 1) - 1;
 
-		if (_vm->_vidPlayer->getCurrentFrame(obj.videoSlot - 1) >= 255) {
+		if ((int32)_vm->_vidPlayer->getCurrentFrame(obj.videoSlot - 1) >= 255) {
 			// Allow for object videos with more than 255 frames, although the
 			// object frame counter is just a byte.
